feat: migrate useLocalStorage from Vite to Next.js
