+++ src/app/hooks/useLocalStorage.ts
@@ -1,0 +1,26 @@
+import { useEffect, useState } from 'react';
+
+export const useLocalStorage = (key: string, initialValue: string) => {
+  const [storedValue, setStoredValue] = useState(() => {
+    try {
+      if (typeof window === 'undefined') {
+        return initialValue;
+      }
+      const value = window.localStorage.getItem(key);
+      return value !== null ? value : initialValue;
+    } catch (error) {
+      console.warn(`Could not read localStorage key "${key}"`, error);
+      return initialValue;
+    }
+  });
+
+  useEffect(() => {
+    try {
+      window.localStorage.setItem(key, storedValue.trim());
+    } catch (error) {
+      console.warn(`Could not write localStorage key "${key}"`, error);
+    }
+  }, [key, storedValue]);
+
+  return [storedValue, setStoredValue] as const;
+};
